requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
pdf2image
pytesseract
